backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
sqlalchemy>=1.4
alembic>=1.12.0
python-multipart>=0.0.6
bcrypt>=4.0.0,<5.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pydantic>=1.10.13
pydantic-settings>=2.0.3
email-validator>=1.3.1
reportlab>=4.0.0
Pillow>=10.0.0
slowapi>=0.1.9
